Moved AuthModel to proper folder
